Prmdr cypress reports (#167)
Add Cypress reporting feature
