+++ app/.gitignore
@@ -7,6 +7,8 @@
 
 # testing
 /coverage
+/cypress/results
+/cypress/screenshots
 
 # production
 /build
