+++ app/package.json
@@ -13,7 +13,11 @@
         "build-storybook": "storybook build",
         "format": "prettier --write \"**/*.+(ts|tsx|js|scss|json|css|md)\"",
         "prepare": "cd ../ && husky install ./app/.husky",
-        "cypress": "cypress open --env CYPRESS_BASE_URL=http://localhost:3000/,CYPRESS_RUN_AS_SMOKETEST=false"
+        "cypress": "cypress open --env CYPRESS_BASE_URL=http://localhost:3000/,CYPRESS_RUN_AS_SMOKETEST=false",
+        "cypress-report-run" : "cypress run --reporter mochawesome --env CYPRESS_BASE_URL=http://localhost:3000/,CYPRESS_RUN_AS_SMOKETEST=false",
+        "cypress-report-merge": "npx mochawesome-merge \"cypress/results/*.json\" > mochawesomemerged.json", 
+        "cypress-report-generate": "npx marge mochawesomemerged.json",
+        "cypress-report": "rm -r cypress/results/* && npm run cypress-report-run && npm run cypress-report-merge && npm run cypress-report-generate"
     },
     "dependencies": {
         "@testing-library/jest-dom": "^5.17.0",
@@ -66,6 +70,8 @@
         "govuk-frontend": "^4.7.0",
         "husky": "^8.0.3",
         "lint-staged": "^14.0.1",
+        "mochawesome": "^7.1.3",
+        "mochawesome-merge": "^4.3.0",
         "prettier": "^3.0.3",
         "prop-types": "^15.8.1",
         "storybook": "^7.4.0",
